♻️ Extract ComposeWrapperWebView with composables
